requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
langchain-community
python-dotenv
sqlalchemy
pydantic
aiohttp
pillow
requests
python-multipart
aiofiles
langchain-google-genai
google-generativeai
httpx
